adapt and write unit tests for statistics page in admin area
