+++ admin/src/components/NavBar.spec.js
@@ -49,28 +49,36 @@ describe('NavBar', () => {
     it('has a link to overview', () => {
       expect(wrapper.findAll('.nav-item').at(0).find('a').attributes('href')).toBe('/')
     })
+
     it('has a link to /user', () => {
       expect(wrapper.findAll('.nav-item').at(1).find('a').attributes('href')).toBe('/user')
     })
+
     it('has a link to /creation', () => {
       expect(wrapper.findAll('.nav-item').at(2).find('a').attributes('href')).toBe('/creation')
     })
+
     it('has a link to /creation-confirm', () => {
       expect(wrapper.findAll('.nav-item').at(3).find('a').attributes('href')).toBe(
         '/creation-confirm',
       )
     })
+
     it('has a link to /contribution-links', () => {
       expect(wrapper.findAll('.nav-item').at(4).find('a').attributes('href')).toBe(
         '/contribution-links',
       )
+    })
+
+    it('has a link to /statistic', () => {
+      expect(wrapper.findAll('.nav-item').at(5).find('a').attributes('href')).toBe('/statistic')
     })
   })
 
   describe('wallet', () => {
     const assignLocationSpy = jest.fn()
     beforeEach(async () => {
-      await wrapper.findAll('.nav-item').at(5).find('a').trigger('click')
+      await wrapper.findAll('.nav-item').at(6).find('a').trigger('click')
     })
 
     it.skip('changes window location to wallet', () => {
@@ -89,7 +97,7 @@ describe('NavBar', () => {
       window.location = {
         assign: windowLocationMock,
       }
-      await wrapper.findAll('.nav-item').at(6).find('a').trigger('click')
+      await wrapper.findAll('.nav-item').at(7).find('a').trigger('click')
     })
 
     it('redirects to /logout', () => {
