+++ admin/src/pages/Overview.spec.js
@@ -1,6 +1,5 @@
 import { mount } from '@vue/test-utils'
 import Overview from './Overview.vue'
-import { communityStatistics } from '@/graphql/communityStatistics.js'
 import { listUnconfirmedContributions } from '@/graphql/listUnconfirmedContributions.js'
 
 const localVue = global.localVue
@@ -20,19 +19,6 @@ const apolloQueryMock = jest
           pending: true,
         },
       ],
-    },
-  })
-  .mockResolvedValueOnce({
-    data: {
-      communityStatistics: {
-        totalUsers: 3113,
-        activeUsers: 1057,
-        deletedUsers: 35,
-        totalGradidoCreated: '4083774.05000000000000000000',
-        totalGradidoDecayed: '-1062639.13634129622923372197',
-        totalGradidoAvailable: '2513565.869444365732411569',
-        totalGradidoUnbookedDecayed: '-500474.6738366222166261272',
-      },
     },
   })
   .mockResolvedValue({
@@ -84,14 +70,6 @@ describe('Overview', () => {
       expect(apolloQueryMock).toBeCalledWith(
         expect.objectContaining({
           query: listUnconfirmedContributions,
-        }),
-      )
-    })
-
-    it('calls communityStatistics', () => {
-      expect(apolloQueryMock).toBeCalledWith(
-        expect.objectContaining({
-          query: communityStatistics,
         }),
       )
     })
